+++ plugin/src/init.server.lua
@@ -20,16 +20,16 @@ local SettingsButton =
 
 --#region Settings
 
-local AnalyticsService = game:GetService("AnalyticsService")
+local TestService = game:GetService("TestService")
 
 local Settings = nil :: any
 
-local SettingsModule = AnalyticsService:FindFirstChild("LuauLSP_Settings") :: ModuleScript
+local SettingsModule = TestService:FindFirstChild("LuauLSP_Settings") :: ModuleScript
 if not SettingsModule then
 	SettingsModule = (script :: any).DefaultSettings:Clone()
 	assert(SettingsModule, "Luau Typechecking")
 	SettingsModule.Name = "LuauLSP_Settings"
-	SettingsModule.Parent = AnalyticsService
+	SettingsModule.Parent = TestService
 end
 assert(SettingsModule, "failed to create settings module")